requirements.txt
aniso8601==9.0.1
arabic-reshaper==2.1.3
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-RESTful==0.3.9
fpdf2==2.4.5
future==0.18.2
google-auth==2.3.2
google-auth-oauthlib==0.4.6
gspread==4.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.3
oauthlib==3.1.1
pandas==1.3.4
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.7
Werkzeug==2.0.2
